% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{e}
\alias{e}
\title{Estimate treatment mechanism conditional on mediators}
\usage{
e(a, m, w)
}
\arguments{
\item{a}{...}

\item{m}{...}

\item{w}{...}
}
\description{
Estimate treatment mechanism conditional on mediators
}
