% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim.R
\name{estimatheta}
\alias{estimatheta}
\title{Stochastic Mediation Effect of Contrasts Under Intermediate Confounding}
\usage{
estimatheta(data, contrast, weights, candidatesg, candidatese, candidatesm,
  candidatesq, candidatesr, candidatesu, candidatesv, nfolds,
  family.outcome)
}
\arguments{
\item{data}{...}

\item{contrast}{...}

\item{weights}{...}

\item{candidatesg}{...}

\item{candidatese}{...}

\item{candidatesm}{...}

\item{candidatesq}{...}

\item{candidatesr}{...}

\item{candidatesu}{...}

\item{candidatesv}{...}

\item{nfolds}{...}

\item{family.outcome}{...}
}
\description{
Stochastic Mediation Effect of Contrasts Under Intermediate Confounding
}
