% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intu}
\alias{intu}
\title{Estimate integral of nuisance function U over Z}
\usage{
intu(w, aprime, astar)
}
\arguments{
\item{w}{...}

\item{aprime}{...}

\item{astar}{...}
}
\description{
Estimate integral of nuisance function U over Z
}
