% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intv}
\alias{intv}
\title{Estimate integral of nuisance function V over Z}
\usage{
intv(m, w, aprime)
}
\arguments{
\item{m}{...}

\item{w}{...}

\item{aprime}{...}
}
\description{
Estimate integral of nuisance function V over Z
}
