% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim.R
\name{mediation}
\alias{mediation}
\title{Stochastic Direct/Indirect Effects Under Intermediate Confounding}
\usage{
mediation(data, weights, candidatesg, candidatese, candidatesm,
  candidatesq, candidatesr, candidatesu, candidatesv, nfolds,
  family.outcome)
}
\arguments{
\item{data}{...}

\item{weights}{...}

\item{candidatesg}{...}

\item{candidatese}{...}

\item{candidatesm}{...}

\item{candidatesq}{...}

\item{candidatesr}{...}

\item{candidatesu}{...}

\item{candidatesv}{...}

\item{nfolds}{...}

\item{family.outcome}{...}
}
\description{
Stochastic Direct/Indirect Effects Under Intermediate Confounding
}
