% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_funs.R
\name{mySL}
\alias{mySL}
\title{Custom Super Learner}
\usage{
mySL(Y, X, newX = NULL, family = stats::gaussian(), SL.library,
  method = "method.NNLS", id = NULL, verbose = FALSE,
  control = list(), cvControl = list(), obsWeights = NULL,
  env = parent.frame(), validRows)
}
\arguments{
\item{Y}{...}

\item{X}{...}

\item{newX}{...}

\item{family}{...}

\item{SL.library}{...}

\item{method}{...}

\item{id}{...}

\item{verbose}{...}

\item{control}{...}

\item{cvControl}{...}

\item{obsWeights}{...}

\item{env}{...}

\item{validRows}{...}
}
\description{
Custom Super Learner
}
