% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pmaw}
\alias{pmaw}
\title{Estimate mediation under intervention on intermediate confounder}
\usage{
pmaw(m, a, w)
}
\arguments{
\item{m}{...}

\item{a}{...}

\item{w}{...}
}
\description{
Estimate mediation under intervention on intermediate confounder
}
