% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pmw}
\alias{pmw}
\title{Estimate mediation under joint intervention on treatment and confounder}
\usage{
pmw(m, w)
}
\arguments{
\item{m}{...}

\item{w}{...}
}
\description{
Estimate mediation under joint intervention on treatment and confounder
}
