% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r}
\alias{r}
\title{Estimate re-parameterized confounding mechanism conditional on mediators}
\usage{
r(z, a, m, w)
}
\arguments{
\item{z}{...}

\item{a}{...}

\item{m}{...}

\item{w}{...}
}
\description{
Estimate re-parameterized confounding mechanism conditional on mediators
}
