% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_funs.R
\name{screen.corRank50}
\alias{screen.corRank50}
\title{Screening via correlation ranking}
\usage{
screen.corRank50(Y, X, family, obsWeights, id, method = "pearson",
  rank = 50, ...)
}
\arguments{
\item{Y}{...}

\item{X}{...}

\item{family}{...}

\item{obsWeights}{...}

\item{id}{...}

\item{method}{...}

\item{rank}{...}

\item{...}{...}
}
\description{
Screening via correlation ranking
}
