% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_funs.R
\name{screen.ttest50}
\alias{screen.ttest50}
\title{Screening via t-test}
\usage{
screen.ttest50(Y, X, family, obsWeights, id, rank = 50, ...)
}
\arguments{
\item{Y}{...}

\item{X}{...}

\item{family}{...}

\item{obsWeights}{...}

\item{id}{...}

\item{rank}{...}

\item{...}{...}
}
\description{
Screening via t-test
}
