% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_funs.R
\name{SL.caretXGB}
\alias{SL.caretXGB}
\title{Custom Super Learner Wrapper for Caret-based xgboost}
\usage{
SL.caretXGB(Y, X, newX, family, obsWeights, ...)
}
\arguments{
\item{Y}{...}

\item{X}{...}

\item{newX}{...}

\item{family}{...}

\item{obsWeights}{...}

\item{...}{...}
}
\description{
Custom Super Learner Wrapper for Caret-based xgboost
}
