% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{u}
\alias{u}
\title{Estimate nuisance function U}
\usage{
u(z, w, aprime, astar)
}
\arguments{
\item{z}{...}

\item{w}{...}

\item{aprime}{...}

\item{astar}{...}
}
\description{
Estimate nuisance function U
}
